# src/main/webui/src/api/model/connectionCreationRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * filature API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */

export interface ConnectionCreationRequest {
  sourceConfigurationId: number;
  targetConfigurationId: number;
  /** @pattern \S */
  targetUploadPath: string;
}