# src/main/webui/src/api/model/connectionSummary.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * filature API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */
import type { SourceSummary } from "./sourceSummary";
import type { TargetSummary } from "./targetSummary";

export interface ConnectionSummary {
  id?: number;
  userId?: string;
  source?: SourceSummary;
  target?: TargetSummary;
  targetUploadPath?: string;
}