# src/main/webui/src/api/model/enercoopSourceSummary.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * filature API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */

export interface EnercoopSourceSummary {
  id?: number;
  name?: string;
  username?: string;
}