# src/main/webui/src/api/model/nextcloudTargetCreationRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * filature API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */

export interface NextcloudTargetCreationRequest {
  /** @pattern \S */
  username: string;
  /** @pattern \S */
  password: string;
  /** @pattern \S */
  url: string;
}