# src/main/webui/src/api/model/synchronizationStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * filature API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */

export type SynchronizationStatus =
  (typeof SynchronizationStatus)[keyof typeof SynchronizationStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SynchronizationStatus = {
  SUCCESS: "SUCCESS",
  FAILURE: "FAILURE",
} as const;