# src/main/webui/src/api/model/synchronizationSummary.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * filature API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */
import type { SynchronizationStatus } from "./synchronizationStatus";
import type { LocalDateTime } from "./localDateTime";

export interface SynchronizationSummary {
  id?: number;
  connectionId?: number;
  status?: SynchronizationStatus;
  error?: string;
  startedAt?: LocalDateTime;
  endedAt?: LocalDateTime;
}